+++ plugins/core/Standard/qCompass/include/ccNoteTool.h
@@ -21,25 +21,22 @@
 #include "ccTool.h"
 #include "ccNote.h"
 
-#include <qinputdialog.h>
-#include <qmainwindow.h>
+#include <QInputDialog>
+#include <QMainWindow>
 
-/*
-Tool used to create notes and associated them with points in a cloud.
-*/
+//! Tool used to create notes and associated them with points in a cloud.
 class ccNoteTool : public ccTool
 {
 public:
 	ccNoteTool();
-	virtual ~ccNoteTool() = default;
 
-	//called when the tool is set to active (for initialization)
-	virtual void toolActivated() override;
+	//! Called when the tool is set to active (for initialization)
+	void toolActivated() override;
 
-	//called when the tool is set to disactive (for cleanup)
-	virtual void toolDisactivated() override;
+	//! Called when the tool is set to disactive (for cleanup)
+	void toolDisactivated() override;
 
-	//called when a point in a point cloud gets picked while this tool is active
+	//! Called when a point in a point cloud gets picked while this tool is active
 	void pointPicked(ccHObject* insertPoint, unsigned itemIdx, ccPointCloud* cloud, const CCVector3& P) override;
 };