+++ plugins/core/Standard/qCompass/include/ccPinchNodeTool.h
@@ -25,23 +25,19 @@
 #include <qinputdialog.h>
 #include <qmainwindow.h>
 
-/*
-Tool used to create PinchNodes.
-*/
-class ccPinchNodeTool :
-	public ccTool
+//! Tool used to create PinchNodes
+class ccPinchNodeTool :	public ccTool
 {
 public:
 	ccPinchNodeTool();
-	virtual ~ccPinchNodeTool() = default;
 
-	//called when the tool is set to active (for initialization)
-	virtual void toolActivated() override;
+	//! Called when the tool is set to active (for initialization)
+	void toolActivated() override;
 
-	//called when the tool is set to disactive (for cleanup)
-	virtual void toolDisactivated() override;
+	//! Called when the tool is set to disactive (for cleanup)
+	void toolDisactivated() override;
 
-	//called when a point in a point cloud gets picked while this tool is active
+	//! Called when a point in a point cloud gets picked while this tool is active
 	void pointPicked(ccHObject* insertPoint, unsigned itemIdx, ccPointCloud* cloud, const CCVector3& P) override;
 };